backend/api/mixins.py
from rest_framework import mixins, viewsets


class RetrieveCreateListViewSet(
    mixins.RetrieveModelMixin,
    mixins.CreateModelMixin,
    mixins.ListModelMixin,
    viewsets.GenericViewSet
):
    pass
